// src/generated-client/models/BeansPostRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Coffee Addiction API
 * Coffee Addiction API swagger documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BeansPostRequest
 */
export interface BeansPostRequest {
    /**
     * Name of the bean
     * @type {string}
     * @memberof BeansPostRequest
     */
    name?: string;
    /**
     * Country or region of origin
     * @type {string}
     * @memberof BeansPostRequest
     */
    origin?: string;
    /**
     * Level of roast (e.g., light, medium, dark)
     * @type {string}
     * @memberof BeansPostRequest
     */
    roastLevel?: string;
    /**
     * Price per kilogram of coffee beans
     * @type {number}
     * @memberof BeansPostRequest
     */
    pricePerKg?: number;
    /**
     * URL
     * @type {string}
     * @memberof BeansPostRequest
     */
    imageUrl?: string;
    /**
     * Available quantity in stock
     * @type {number}
     * @memberof BeansPostRequest
     */
    stockQuantity?: number;
    /**
     * Detailed description of the coffee bean
     * @type {string}
     * @memberof BeansPostRequest
     */
    description?: string;
}

/**
 * Check if a given object implements the BeansPostRequest interface.
 */
export function instanceOfBeansPostRequest(value: object): value is BeansPostRequest {
    return true;
}

export function BeansPostRequestFromJSON(json: any): BeansPostRequest {
    return BeansPostRequestFromJSONTyped(json, false);
}

export function BeansPostRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): BeansPostRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'origin': json['origin'] == null ? undefined : json['origin'],
        'roastLevel': json['roastLevel'] == null ? undefined : json['roastLevel'],
        'pricePerKg': json['pricePerKg'] == null ? undefined : json['pricePerKg'],
        'imageUrl': json['imageUrl'] == null ? undefined : json['imageUrl'],
        'stockQuantity': json['stockQuantity'] == null ? undefined : json['stockQuantity'],
        'description': json['description'] == null ? undefined : json['description'],
    };
}

export function BeansPostRequestToJSON(json: any): BeansPostRequest {
    return BeansPostRequestToJSONTyped(json, false);
}

export function BeansPostRequestToJSONTyped(value?: BeansPostRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'origin': value['origin'],
        'roastLevel': value['roastLevel'],
        'pricePerKg': value['pricePerKg'],
        'imageUrl': value['imageUrl'],
        'stockQuantity': value['stockQuantity'],
        'description': value['description'],
    };
}